requirements.txt
blinker==1.9.0
click==8.1.7
Faker==33.1.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pip==22.0.2
python-dateutil==2.9.0.post0
setuptools==59.6.0
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
